+++ test/test_helpers/asset_manager_test.rb
@@ -9,11 +9,11 @@ describe GdsApi::TestHelpers::AssetManager do
     GdsApi::AssetManager.new(Plek.find("asset-manager"))
   end
 
-  describe "#stub_asset_manager_receives_an_asset" do
+  describe "#stub_asset_manager_create_asset" do
     describe "when passed a string" do
       it "returns the string as the file url" do
         url = "https://assets.example.com/path/to/asset"
-        stub_asset_manager_receives_an_asset(url)
+        stub_asset_manager_create_asset(url)
         response = stub_asset_manager.create_asset({})
 
         assert_equal url, response["file_url"]
@@ -22,7 +22,7 @@ describe GdsApi::TestHelpers::AssetManager do
 
     describe "when passed no arguments" do
       it "returns a random, yet valid asset manager url" do
-        stub_asset_manager_receives_an_asset
+        stub_asset_manager_create_asset
         response = stub_asset_manager.create_asset({})
 
         url_format = %r{\Ahttp://asset-manager.dev.gov.uk/media/[^/]*/[^/]*\Z}
@@ -30,7 +30,7 @@ describe GdsApi::TestHelpers::AssetManager do
       end
 
       it "returns a different URL each call" do
-        stub_asset_manager_receives_an_asset
+        stub_asset_manager_create_asset
         response1 = stub_asset_manager.create_asset({})
         response2 = stub_asset_manager.create_asset({})
 
@@ -40,7 +40,7 @@ describe GdsApi::TestHelpers::AssetManager do
 
     describe "when passed a hash" do
       it "can specify the id of an asset" do
-        stub_asset_manager_receives_an_asset(id: "123")
+        stub_asset_manager_create_asset(id: "123")
         response = stub_asset_manager.create_asset({})
 
         url_format = %r{\Ahttp://asset-manager.dev.gov.uk/media/123/[^/]*\Z}
@@ -48,7 +48,7 @@ describe GdsApi::TestHelpers::AssetManager do
       end
 
       it "can specify the filename of an asset" do
-        stub_asset_manager_receives_an_asset(filename: "file.ext")
+        stub_asset_manager_create_asset(filename: "file.ext")
         response = stub_asset_manager.create_asset({})
 
         url_format = %r{\Ahttp://asset-manager.dev.gov.uk/media/[^/]*/file.ext\Z}
@@ -56,12 +56,43 @@ describe GdsApi::TestHelpers::AssetManager do
       end
 
       it "can specify both filename and id" do
-        stub_asset_manager_receives_an_asset(id: "123", filename: "file.ext")
+        stub_asset_manager_create_asset(id: "123", filename: "file.ext")
         response = stub_asset_manager.create_asset({})
 
         url_format = %r{\Ahttp://asset-manager.dev.gov.uk/media/123/file.ext\Z}
         assert_match url_format, response["file_url"]
       end
+    end
+
+    describe "when passed a response body" do
+      it "returns the body in the response" do
+        url = "https://assets.example.com/path/to/asset"
+        response_body = { "key": "value" }
+        stub_asset_manager_create_asset(url, response_body)
+        response = stub_asset_manager.create_asset({})
+
+        assert_equal "value", response["key"]
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_receives_an_asset" do
+    it "includes a deprecation warning" do
+      expects(:warn).with(
+        <<~WARNING,
+          stub_asset_manager_receives_an_asset is deprecated and will be removed
+          in a future version of gds-api-adapters. Replace calls to this method with
+          stub_asset_manager_create_asset.
+        WARNING
+      )
+
+      stub_asset_manager_receives_an_asset
+    end
+
+    it "delegates to the new method" do
+      expects(:stub_asset_manager_create_asset).once
+
+      stub_asset_manager_receives_an_asset
     end
   end
 
@@ -96,6 +127,134 @@ describe GdsApi::TestHelpers::AssetManager do
 
       assert_raises GdsApi::HTTPInternalServerError do
         stub_asset_manager.delete_asset(asset_id)
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_has_an_asset" do
+    describe "when the asset endpoint is requested" do
+      it "returns the given attributes" do
+        asset_id = "some-asset-id"
+        body = { key: "value" }
+        stub_asset_manager_has_an_asset(asset_id, body)
+        response = stub_asset_manager.asset(asset_id)
+
+        assert_equal 200, response.code
+        assert_equal body[:key], response["key"]
+      end
+    end
+
+    describe "when the media endpoint is requested" do
+      it "returns the file content" do
+        asset_id = "some-asset-id"
+        body = { key: "value" }
+        filename = "file.pdf"
+        stub_asset_manager_has_an_asset(asset_id, body, filename)
+        response = stub_asset_manager.media(asset_id, filename)
+
+        assert_equal 200, response.code
+        assert_equal "Some file content", response
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_request_to_get_asset" do
+    describe "when the asset endpoint is requested" do
+      it "returns the given attributes" do
+        asset_id = "some-asset-id"
+        body = { key: "value" }
+        stub_asset_manager_request_to_get_asset(asset_id, body)
+        response = stub_asset_manager.asset(asset_id)
+
+        assert_equal 200, response.code
+        assert_equal body[:key], response["key"]
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_request_to_asset_media" do
+    describe "when the media endpoint is requested" do
+      it "returns the file content" do
+        asset_id = "some-asset-id"
+        filename = "file.pdf"
+        stub_asset_manager_request_to_asset_media(asset_id, filename)
+        response = stub_asset_manager.media(asset_id, filename)
+
+        assert_equal 200, response.code
+        assert_equal "Some file content", response
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_create_asset_too_large" do
+    describe "when the endpoint is requested" do
+      it "raises GdsApi::HTTPPayloadTooLarge" do
+        stub_asset_manager_create_asset_too_large
+
+        assert_raises GdsApi::HTTPPayloadTooLarge do
+          stub_asset_manager.create_asset({})
+        end
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_create_asset_unprocessable" do
+    describe "when the endpoint is requested" do
+      it "raises GdsApi::HTTPUnprocessableEntity" do
+        stub_asset_manager_create_asset_unprocessable
+
+        assert_raises GdsApi::HTTPUnprocessableEntity do
+          stub_asset_manager.create_asset({})
+        end
+      end
+
+      it "includes the body when provided" do
+        stub_asset_manager_create_asset_unprocessable("Some output")
+
+        error = assert_raises GdsApi::HTTPUnprocessableEntity do
+          stub_asset_manager.create_asset({})
+        end
+
+        assert_includes error.message, "Some output"
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_update_asset_forbidden" do
+    describe "when the endpoint is requested with the given ID" do
+      it "raises GdsApi::HTTPForbidden" do
+        asset_id = "some-id"
+        stub_asset_manager_update_asset_forbidden(asset_id)
+
+        assert_raises GdsApi::HTTPForbidden do
+          stub_asset_manager.update_asset(asset_id, {})
+        end
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_update_asset_not_found" do
+    describe "when the endpoint is requested with the given ID" do
+      it "raises GdsApi::HTTPNotFound" do
+        asset_id = "some-id"
+        stub_asset_manager_update_asset_not_found(asset_id)
+
+        assert_raises GdsApi::HTTPNotFound do
+          stub_asset_manager.update_asset(asset_id, {})
+        end
+      end
+    end
+  end
+
+  describe "#stub_asset_manager_get_asset_forbidden" do
+    describe "when the endpoint is requested with the given ID" do
+      it "raises GdsApi::HTTPForbidden" do
+        asset_id = "some-id"
+        stub_asset_manager_get_asset_forbidden(asset_id)
+
+        assert_raises GdsApi::HTTPForbidden do
+          stub_asset_manager.asset(asset_id)
+        end
       end
     end
   end